# [GSoC 2018] Multistream API for vocabulary building in *2vec