… gensim/models/fasttext.py
… gensim/models/word2vec.py
… gensim/scripts/benchmark_vocab.py
… gensim/test/test_doc2vec.py
… gensim/test/test_fasttext.py
… gensim/test/test_utils.py
… gensim/test/test_word2vec.py
… gensim/utils.py