This PR introduces __multi-stream API__ for building vocabularies in __*2Vec models__. 

# Main idea

Training *2vec models involves two main steps:

1. One pass over your training corpus to build (and filter) the vocabulary
2. Train the model

This pull request optimizes the __first step__ (dictionary building).

The main idea is to parallelize the `scan_vocab` method, by allowing users to pass _several streams of data_ (i.e. several `LineSentence` iterators for different files). Under the hood, the implementation uses `multiprocessing` python module creating `multiprocessing.Pool(processes=min(workers, len(input_streams)))` to __read and process all input streams in parallel__.

# Why use it?

This new feature substantially reduces the vocabulary building time, especially if you have complex data iterators (more CPU-bound, i.e. reading a line from a file, then doing complex preprocessing, removing stop words, lemmatizing, stemming etc).

In our benchmarks below, __multi-stream mode__ produces up to __3x boost__ in the vocabulary building step. This is significant because this step can take hours for larger corpora (e.g. on https://dumps.wikimedia.org/enwiki/20180320/ :)

The more CPU-complex data iterators you use, the more speedup you will get. For example, I've experimented with building vocab for `Word2Vec` model in two setups, **without preprocessing** and **with preprocessing**:

1. **Without preprocessing**. Iterate with `gensim.models.word2vec.LineSentence`, which just reads a line from disk and splits it on whitespace. I ran this benchmark on already preprocessed dataset. Using multi-stream resulted in **2.7x speedup**.

| # workers | total time, secs |
|:---------:|:-:|
|     1     |      249.51     |
|     2     |      159.87     |
|     3     |      126.47     |
|     4     |      109.90     |
|     5     |      96.79      |
|     8     |      88.74      |
|     10    |      93.00      |

2. **With preprocessing**. Iterate over lines and preprocess each one with `gensim.parsing.preprocessing.preprocess_string`. In this case, multistream mode has provided **6.9x speedup**.

| # workers | total time, secs |
|:---------:|:-:|
|     1     |      4398.54     |
|     2     |      2252.73     |
|     3     |      1596.28     |
|     4     |      1157.02     |
|     5     |      926.66      |
|     8     |      704.07      |
|     10    |      638.45      |

# How to use it?

It's really simple to start using the multi-stream mode. All you need is to have multiple document streams (training corpora, e.g. files on disk) you want to train on, and pass them as a list into the new `input_streams` constructor parameter. `input_streams` supersedes the  old `sentences` parameter which is now equivalent to `input_streams = [sentences]`:

### Old single stream approach

1. Build vocabulary and train at the same time

```python
sentences = LineSentence('myfile.txt')
model = Word2Vec(sentences, workers=10, ...other params...)
```

2. Initialize model, build vocab and train as separate steps

```python
sentences = LineSentence('myfile.txt')
model = Word2Vec(workers=10, ...other params...)
model.build_vocab(sentences)
model.train(sentences, ...train params...)
```

### New multi-stream approach

1. Build vocab and train at the same time (parallelized internally)

```python
input_streams = [LineSentence('myfile1.txt'), LineSentence('myfile2.txt'), ...]
model = Word2Vec(input_streams=input_streams, workers=10, ...other params...)
```

2. Initialize model, build vocab in parallel and train as separate steps

```python
input_streams = [LineSentence('myfile1.txt'), LineSentence('myfile2.txt'), ...]
model = Word2Vec(workers=10, ...other params...)
model.build_vocab(input_streams=input_streams)  # Optionally, you can pass `workers` param here which differs from the one passed in constructor.
model.train(input_streams=input_streams, ...train params...)
```

The examples above used `Word2Vec`, but `Doc2Vec` and `FastText` work analogously.