@persiyanov 2018-06-04
```
* 2018-06-04 14:11:10,727 : INFO : Model = word2vec	Workers = 1	Vocab time = 220.48 secs
*  2018-06-04 14:13:39,131 : INFO : Model = word2vec	Workers = 2	Vocab time = 147.40 secs
*  2018-06-04 14:15:44,110 : INFO : Model = word2vec	Workers = 3	Vocab time = 123.86 secs
*  2018-06-04 14:17:30,654 : INFO : Model = word2vec	Workers = 4	Vocab time = 105.28 secs
*  2018-06-04 14:19:00,277 : INFO : Model = word2vec	Workers = 5	Vocab time = 88.51 secs
*  2018-06-04 14:20:24,006 : INFO : Model = word2vec	Workers = 8	Vocab time = 82.50 secs
*  2018-06-04 14:21:43,277 : INFO : Model = word2vec	Workers = 10	Vocab time = 77.99 secs

*  2018-06-04 15:09:04,340 : INFO : Model = doc2vec	Workers = 1	Vocab time = 223.81 secs
*  2018-06-04 15:11:39,150 : INFO : Model = doc2vec	Workers = 2	Vocab time = 153.86 secs
*  2018-06-04 15:13:51,637 : INFO : Model = doc2vec	Workers = 3	Vocab time = 131.49 secs
*  2018-06-04 15:15:43,956 : INFO : Model = doc2vec	Workers = 4	Vocab time = 111.28 secs
*  2018-06-04 15:17:23,360 : INFO : Model = doc2vec	Workers = 5	Vocab time = 98.30 secs
*  2018-06-04 15:18:55,534 : INFO : Model = doc2vec	Workers = 8	Vocab time = 91.04 secs
*  2018-06-04 15:20:22,721 : INFO : Model = doc2vec	Workers = 10	Vocab time = 86.03 secs

* 2018-06-04 14:43:46,260 : INFO : Model = fasttext	Workers = 1	Vocab time = 309.86 secs
*  2018-06-04 14:47:52,583 : INFO : Model = fasttext	Workers = 2	Vocab time = 244.43 secs
*  2018-06-04 14:51:32,450 : INFO : Model = fasttext	Workers = 3	Vocab time = 217.69 secs
*  2018-06-04 14:54:55,467 : INFO : Model = fasttext	Workers = 4	Vocab time = 201.10 secs
*  2018-06-04 14:58:06,800 : INFO : Model = fasttext	Workers = 5	Vocab time = 189.07 secs
*  2018-06-04 15:01:08,160 : INFO : Model = fasttext	Workers = 8	Vocab time = 179.37 secs
*  2018-06-04 15:04:05,421 : INFO : Model = fasttext	Workers = 10	Vocab time = 175.31 secs
```