@persiyanov 2018-06-05
@gojomo Hi! I'd like to ask for your advice. There is one bug in the code in this PR related to `TaggedLineSentence` and tags produced by it. If a user wants to use multistream + `TaggedLineSentence`, then there will be several documents with the same tags (any two documents on the same line in different files will have the same tags, because of `TaggedLineSentence` logic). Could you suggest how to fix this problem?