@gojomo 2018-06-05
Clearly incrementing-serial-numbers-per-line can't be used if you have multiple files that all start with line 0. You could require tags to be specified inside the files – as an extra field/fields each line. If the files are considered in a stable order, and the count of texts (lines) in each can be pre-determined (perhaps with another extra pass over all data), each could be given a unique starting-id. Text tags could also be a deterministic function of the hash of the contained text – for additional per-line overhead, and the extra cost of keeping those (now much longer) tags as string keys.