@menshikh-iv 2018-06-06
@persiyanov great work! Please add tests & merge fresh `master` to your branch (just in case).

@gojomo problem mostly defined by user-behavior, i.e. if I (as gensim-user) heard about new multistream feature and want to try it with d2v - I'll use something like  
`[TaggedLineDocument(..), TaggedLineDocument(..)], multistream=True` (that's definitely will not work as we expected in this case).