@persiyanov 2018-06-08
hooray! Tests are passing now. I will update the description of the PR in a proper way (what's the feature, how it helps, how to use it etc.)