@persiyanov 2018-06-13
@gojomo @jayantj @piskvorky Please, take one more look. I've added final vocabulary pruning