@gojomo 2018-06-14
If it works, great, but I'm a little uncomfortable with the manner-of-activation: the `multistream` boolean toggle, and thus differing if-this-type-then overloading of a single `sentences` parameter, all as new branches inside existing code.

Ultimately, even if not right away, I'd prefer a mechanism where people opting-into an alternative-vocabulary-building-optimization use some other explicit code path, not complicating existing interfaces, then drop a completed vocabulary-object into the Word2Vec/etc model. Then the simple/legacy path stays simple, the new alternative is clearly opted-into, and the point-of-entanglement is limited to a single injection of what the main model needs to proceed – which also allows the main model to be indifferent as to the specifics of how its individual  steps were completed. That could also allow further experimentation with other vocab-building approaches, without adding more switches-and-branches to a single entry point every time. (That there's only a 2.5X-3X speedup in moving from 1 to 10 threads makes me think further improvements may be possible.)