@persiyanov 2018-06-14
@gojomo Could you give some examples on what you mean by `opting-into an alternative-vocabulary-building-optimization use some other explicit code path...`? Something like `vocab = gensim.models.word2vec.build_vocab_multistream_mode(input_streams, <vocab params>)` and then `model.vocab = vocab`? Or something different?

Why do you think the interface I proposed is complicated? It's just toggling one flag (`multistream=True`) and start passing _multiple streams_ instead of single stream. This looks very clear and simple, isn't it?

Is it okay to merge this PR in its current state (after addressing @piskvorky comments) and possibly change the interfaces and API in the future (according to new research & results in __multistream training__ direction) if needed?