@gojomo 2018-06-15
@persiyanov Yes, your example utility function `build_vocab_multistream_mode()` that returns a usable vocabulary-object is more along the lines of what I'd prefer. 

The switches-and-branches approach works, and if there's just one alt mode, not too complex. But, it's expanding some (already-gigantic) method-parameter lists, and some already-twisty methods - making them harder to read/maintain/expand in the future. 

If just a switch, the docs for this mode will be interleaved with long lists of other options – which may tend to tempt those who don't quite need this, but shortchange those who do want full descriptions. Having the docs on a specialized function or class may be more clear – and better able to discuss subtleties like the potentially different memory requirements of this approach, or its sensitivity to the various parallel streams being of roughly-equal size.

Also, adding new switches for new approaches in the future, which might or might not be compatible with each other, will be harder than having a separate, swappable step that hides its own details and just leaves things in a proper state. For example, there's another effort in #1962 to offer the approximate-counting `Bounter` as a vocab-surveying option... but it *also* adds a toggle-on-parameter and bunch of if-thens to nearby code. As edits to a single class, these would be hard to merge; as truly standalone paths-to-a-working-vocabulary-object, they'd be easily swappable as either (1) a policy/implementation option supplied to the main class; or (2) a separate step whose output is just provided to the mail class when done. (You can see I've raised similar modularity concerns at that PR.)

The preexisting and not-so-flexible tangle of the *2Vec initialization steps long predates your work, so it may be outside your project to fully fix it, but I'm wary of making the tangle any thicker.