@persiyanov 2018-06-19
@piskvorky pls, take a look. I've addressed your comments