@persiyanov 2018-07-11
@piskvorky Are you OK with this PR? If so, let's merge it. I need this changes for my next pull request related to multistream training.