@piskvorky 2018-07-11
Yeah, I think so. I don't see any outstanding unresolved comments/issues. @menshikh-iv ?